# replica: remove rwlock for protecting iteration over storage group map