… replica/table.cc
… test/topology_experimental_raft/test_tablets.py