@raphaelsc 2024-06-24
v2:
- gets rid of fragile intrusive list (another source of truth which can result in inconsistency).
- added documentation to explain scenarios and how each is handled without rwlock
- improved storage_group::stop() by launching close of gate first, and waiting for it after stopping compaction groups (to avoid waiting on ongoing compactions)