@raphaelsc 2024-06-26
v3:
- fixed scylla-gdb.py
- added another test to increase confidence
- hardened it further by holding guard of compaction groups in parallel_foreach_compaction_group