@bhalevy 2024-07-01
@raphaelsc we must get perf-simple-query results (both read and write) before and after this change
to make sure there is no regressions (or even better, show some progression ;-))