@raphaelsc 2024-07-01
> @raphaelsc we must get perf-simple-query results (both read and write) before and after this change to make sure there is no regressions (or even better, show some progression ;-))

I wouldn't really expect any improvement, other than not stalling topology changes anymore, which is already a huge deal. I will be happy if first result shows there is no regression :-)