@raphaelsc 2024-07-03
v4:
- get rid of table::compaction_groups()
- change table::active_memtables() to table::for_each_active_memtable().
- use coroutine::as_future() when closing gate
- parallelize iteration at compaction group level too
- not lose backward compatibility in scylla-gdb.py
- perf regression test included in cover letter