@raphaelsc 2024-07-09
v5:
- rebased
- annotated 'storage_group::compaction_groups() const' with noexcept
- went back to previous approach of closing gate in stop()