@scylladb-promoter 2024-07-10
#### :red_circle: CI State: FAILURE

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10300/)
:white_check_mark: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10300/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: boost/commitlog_test
:small_blue_diamond: boost/memtable_test
:small_blue_diamond: boost/mutation_test
:small_blue_diamond: topology_experimental_raft/test_tablets
:white_check_mark: - [Container Test](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10300/)
:x: - [dtest](https://jenkins.scylladb.com/job/scylla-master/job/gating-dtest-release/10556/)
:white_check_mark: - [dtest with topology changes](https://jenkins.scylladb.com/job/scylla-master/job/gating-dtest-release-with-consistent-topology-changes/2531/)
:white_check_mark: - [Unit Tests](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10300/)
#### Failed Tests (1/39088):
- [test_concurrent_restart](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/10300/testReport/junit/nodetool_additional_test/TestNodetool/Tests___dtest___test_concurrent_restart) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_concurrent_restart)
#### Build Details:
- Duration: 14 hr
- Builder: i-011c41b6a7e3d625a (m5ad.8xlarge)