@raphaelsc 2024-07-10
> * [test_concurrent_restart](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/10300/testReport/junit/nodetool_additional_test/TestNodetool/Tests___dtest___test_concurrent_restart) [🔍](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_concurrent_restart)

that's known issue: https://github.com/scylladb/scylladb/issues/15192