@scylladb-promoter 2024-07-11
#### :red_circle: CI State: FAILURE

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10340/)
:white_check_mark: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10340/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: boost/commitlog_test
:small_blue_diamond: boost/memtable_test
:small_blue_diamond: boost/mutation_test
:small_blue_diamond: topology_experimental_raft/test_tablets
:white_check_mark: - [Container Test](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10340/)
:white_check_mark: - [dtest](https://jenkins.scylladb.com/job/scylla-master/job/gating-dtest-release/10573/)
:white_check_mark: - [dtest with topology changes](https://jenkins.scylladb.com/job/scylla-master/job/gating-dtest-release-with-consistent-topology-changes/2551/)
:x: - [Unit Tests](https://jenkins.scylladb.com/job/scylla-master/job/scylla-ci/10340/)
#### Failed Tests (1/39333):
- [test_tablet_drain_failure_during_decommission.dev.5](https://jenkins.scylladb.com//job/scylla-master/job/scylla-ci/10340/testReport/junit/%28root%29/test_topology_failure_recovery/Tests___Unit_Tests___test_tablet_drain_failure_during_decommission_dev_5) [:mag:](https://github.com/scylladb/scylladb/issues?q=is:issue+is:open+test_tablet_drain_failure_during_decommission.dev.5)
#### Build Details:
- Duration: 6 hr 24 min
- Builder: spider6.cloudius-systems.com