@raphaelsc 2024-07-11
> * test_tablet_drain_failure_during_decommission

opened https://github.com/scylladb/scylladb/issues/19707, failure is unrelated.